requirements.txt
langgraph==1.0.4
langchain-openai==1.1.1
langchain-core==1.1.3
qdrant-client==1.16.1
python-dotenv==1.2.1
langchain-huggingface==1.1.0
transformers==4.57.3
torch==2.9.1
accelerate==1.12.0
pypdf==6.4.1
fastembed==0.7.4
gradio==6.1.0
plotly==6.5.0
pandas==2.3.3
langchain-groq==1.1.0
xgboost==3.1.2
tweet-preprocessor==0.6.0
